💄 Update all SASS files to current standards
I noticed that I was seeing hundreds of errors in the `jekyll build` and
I decided to fix that. Turns out I'm still using basically-basic as a
gem within jekyll and until I fully rip it out and copy a ton more files
or just switch to some maintained theme in the future that I had to
patch a lot of files.

Most of the fixes involved @import --> @use, or importing SASS modules
like `sass:meta`. The final note here is there were two side effects:

- The main.scss file is being ignored and the one in the gem is used
  instead. So I just changed the local Gemfile location for this and
  left the instructions in my `main.scss`.
- The padding isn't quite as robust as it was before. Not sure why that
  is happening, but I was unable to debug it and adding just a simple
  default 2rem left-padding pretty much got what I wanted anyway.